requirements.txt
torch>=2.0.0
accelerate>=0.20.0
transformers>=4.30.0
diffusers>=0.15.0
albumentations>=1.3.0
opencv-python>=4.7.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=9.5.0
tqdm>=4.65.0
minio>=7.1.0
scikit-image>=0.20.0
